+++ src/Resources/Schema.php
@@ -2,6 +2,8 @@
 namespace DreamFactory\Core\SqlDb\Resources;
 
 use DreamFactory\Core\Database\Schema\ColumnSchema;
+use DreamFactory\Core\Database\Schema\RelationSchema;
+use DreamFactory\Core\Database\Schema\TableSchema;
 use DreamFactory\Core\Exceptions\BadRequestException;
 use DreamFactory\Core\Exceptions\InternalServerErrorException;
 use DreamFactory\Core\Exceptions\NotFoundException;
@@ -74,6 +76,27 @@ class Schema extends BaseDbSchemaResource
     /**
      * {@inheritdoc}
      */
+    public function describeRelationship($table, $relationship, $refresh = false)
+    {
+        if (empty($table)) {
+            throw new BadRequestException('Table name can not be empty.');
+        }
+
+        try {
+            $result = $this->describeTableRelationships($table, $relationship);
+
+            return array_get($result, 0);
+        } catch (RestException $ex) {
+            throw $ex;
+        } catch (\Exception $ex) {
+            throw new InternalServerErrorException("Error describing database table '$table' relationship '$relationship'.\n" .
+                $ex->getMessage(), $ex->getCode());
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function createTables($tables, $check_exist = false, $return_schema = false)
     {
         $tables = static::validateAsArray($tables, null, true, 'There are no table sets in the request.');
@@ -128,13 +151,39 @@ class Schema extends BaseDbSchemaResource
 
         $fields = static::validateAsArray($properties, null, true, 'Bad data format in request.');
 
-        $result = $this->schema->updateFields($table, $fields);
+        $tables = [['name' => $table, 'field' => $fields]];
+        $result = $this->schema->updateSchema($tables, true);
+        $result = array_get(array_get($result, 0, []), 'field', []);
 
         //  Any changes here should refresh cached schema
         $this->refreshCachedTables();
 
         if ($return_schema) {
             return $this->describeField($table, $field);
+        }
+
+        return $result;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function createRelationship($table, $relationship, $properties = [], $check_exist = false, $return_schema = false)
+    {
+        $properties = (is_array($properties) ? $properties : []);
+        $properties['name'] = $relationship;
+
+        $fields = static::validateAsArray($properties, null, true, 'Bad data format in request.');
+
+        $tables = [['name' => $table, 'related' => $fields]];
+        $result = $this->schema->updateSchema($tables, true);
+        $result = array_get(array_get($result, 0, []), 'related', []);
+
+        //  Any changes here should refresh cached schema
+        $this->refreshCachedTables();
+
+        if ($return_schema) {
+            return $this->describeRelationship($table, $relationship);
         }
 
         return $result;
@@ -202,13 +251,43 @@ class Schema extends BaseDbSchemaResource
 
         $fields = static::validateAsArray($properties, null, true, 'Bad data format in request.');
 
-        $result = $this->schema->updateFields($table, $fields, true);
+        $tables = [['name' => $table, 'field' => $fields]];
+        $result = $this->schema->updateSchema($tables, true);
+        $result = array_get(array_get($result, 0, []), 'field', []);
 
         //  Any changes here should refresh cached schema
         $this->refreshCachedTables();
 
         if ($return_schema) {
             return $this->describeField($table, $field);
+        }
+
+        return $result;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function updateRelationship($table, $relationship, $properties = [], $allow_delete_parts = false, $return_schema = false)
+    {
+        if (empty($table)) {
+            throw new BadRequestException('Table name can not be empty.');
+        }
+
+        $properties = (is_array($properties) ? $properties : []);
+        $properties['name'] = $relationship;
+
+        $fields = static::validateAsArray($properties, null, true, 'Bad data format in request.');
+
+        $tables = [['name' => $table, 'related' => $fields]];
+        $result = $this->schema->updateSchema($tables, true);
+        $result = array_get(array_get($result, 0, []), 'related', []);
+
+        //  Any changes here should refresh cached schema
+        $this->refreshCachedTables();
+
+        if ($return_schema) {
+            return $this->describeRelationship($table, $relationship);
         }
 
         return $result;
@@ -266,6 +345,31 @@ class Schema extends BaseDbSchemaResource
     }
 
     /**
+     * {@inheritdoc}
+     */
+    public function deleteRelationship($table, $relationship)
+    {
+        if (empty($table)) {
+            throw new BadRequestException('Table name can not be empty.');
+        }
+
+        // does it already exist
+        if (!$this->doesTableExist($table)) {
+            throw new NotFoundException("A table with name '$table' does not exist in the database.");
+        }
+
+        try {
+            $this->schema->dropRelationship($table, $relationship);
+        } catch (\Exception $ex) {
+            error_log($ex->getMessage());
+            throw $ex;
+        }
+
+        //  Any changes here should refresh cached schema
+        $this->refreshCachedTables();
+    }
+
+    /**
      * @param string $name
      *
      * @throws BadRequestException
@@ -279,6 +383,8 @@ class Schema extends BaseDbSchemaResource
         } else {
             throw new NotFoundException('Table "' . $name . '" does not exist in the database.');
         }
+
+        return $name;
     }
 
     /**
@@ -295,29 +401,29 @@ class Schema extends BaseDbSchemaResource
 
     /**
      * @param string                $table_name
-     * @param null | string | array $field_names
+     * @param null | string | array $fields
      * @param bool                  $refresh
      *
      * @throws NotFoundException
      * @throws InternalServerErrorException
      * @return array
      */
-    public function describeTableFields($table_name, $field_names = null, $refresh = false)
+    public function describeTableFields($table_name, $fields = null, $refresh = false)
     {
         $table = $this->schema->getTable($table_name, $refresh);
         if (!$table) {
             throw new NotFoundException("Table '$table_name' does not exist in the database.");
         }
 
-        if (!empty($field_names)) {
-            $field_names = static::validateAsArray($field_names, ',', true, 'No valid field names given.');
+        if (!empty($fields)) {
+            $fields = static::validateAsArray($fields, ',', true, 'No valid field names given.');
         }
 
         $out = [];
         try {
             /** @var ColumnSchema $column */
             foreach ($table->columns as $column) {
-                if (empty($field_names) || (false !== array_search($column->name, $field_names))) {
+                if (empty($fields) || (false !== array_search($column->name, $fields))) {
                     $out[] = $column->toArray();
                 }
             }
@@ -327,6 +433,46 @@ class Schema extends BaseDbSchemaResource
 
         if (empty($out)) {
             throw new NotFoundException("No requested fields found in table '$table_name'.");
+        }
+
+        return $out;
+    }
+
+    /**
+     * @param string                $table_name
+     * @param null | string | array $relationships
+     * @param bool                  $refresh
+     *
+     * @throws NotFoundException
+     * @throws InternalServerErrorException
+     * @return array
+     */
+    public function describeTableRelationships($table_name, $relationships = null, $refresh = false)
+    {
+        /** @var TableSchema $table */
+        $table = $this->schema->getTable($table_name, $refresh);
+        if (!$table) {
+            throw new NotFoundException("Table '$table_name' does not exist in the database.");
+        }
+
+        if (!empty($relationships)) {
+            $relationships = static::validateAsArray($relationships, ',', true, 'No valid relationship names given.');
+        }
+
+        $out = [];
+        try {
+            /** @var RelationSchema $relation */
+            foreach ($table->relations as $relation) {
+                if (empty($relationships) || (false !== array_search($relation->name, $relationships))) {
+                    $out[] = $relation->toArray();
+                }
+            }
+        } catch (\Exception $ex) {
+            throw new InternalServerErrorException("Failed to query table relationship schema.\n{$ex->getMessage()}");
+        }
+
+        if (empty($out)) {
+            throw new NotFoundException("No requested relationships found in table '$table_name'.");
         }
 
         return $out;