+++ src/Resources/StoredProcedure.php
@@ -195,9 +195,10 @@ class StoredProcedure extends BaseDbResource
     protected function handleGET()
     {
         if (empty($this->resource)) {
-            $names = $this->request->getParameter(ApiOptions::IDS);
+            $payload = $this->request->getPayloadData();
+            $names = array_get($payload, ApiOptions::IDS, $this->request->getParameter(ApiOptions::IDS));
             if (empty($names)) {
-                $names = ResourcesWrapper::unwrapResources($this->request->getPayloadData());
+                $names = ResourcesWrapper::unwrapResources($payload);
             }
 
             if (!empty($names)) {
@@ -294,6 +295,7 @@ class StoredProcedure extends BaseDbResource
     protected function callProcedure()
     {
         $payload = $this->request->getPayloadData();
+        // check payload first, then inline, then URL param
         $params = array_get($payload, 'params', $this->inlineParams);
         if (empty($params)) {
             $params = $this->request->getParameters();
@@ -328,34 +330,26 @@ class StoredProcedure extends BaseDbResource
         }
 
         // convert result field values to types according to schema received
-        if (is_array($schema) && !empty($result)) {
-            if (is_array($result)) {
-                foreach ($result as &$row) {
-                    if (is_array($row)) {
-                        if (isset($row[0])) {
-                            //  Multi-row set, dig a little deeper
-                            foreach ($row as &$sub) {
-                                if (is_array($sub)) {
-                                    foreach ($sub as $key => $value) {
-                                        if (null !== $type = array_get($schema, $key)) {
-                                            $sub[$key] = DataFormatter::formatValue($value, $type);
-                                        }
-                                    }
+        if (is_array($schema) && is_array($result) && !empty($result)) {
+            foreach ($result as $rkey => &$row) {
+                if (is_array($row)) {
+                    //  Multi-row set, dig a little deeper
+                    foreach ($row as $skey => &$sub) {
+                        if (is_array($sub)) {
+                            foreach ($sub as $key => $value) {
+                                if (null !== $type = array_get($schema, $key)) {
+                                    $sub[$key] = DataFormatter::formatValue($value, $type);
                                 }
                             }
                         } else {
-                            foreach ($row as $key => $value) {
-                                if (null !== $type = array_get($schema, $key)) {
-                                    $row[$key] = DataFormatter::formatValue($value, $type);
-                                }
+                            if (null !== $type = array_get($schema, $skey)) {
+                                $row[$skey] = DataFormatter::formatValue($sub, $type);
                             }
                         }
                     }
-                }
-            } else {
-                foreach ($result as $key => $value) {
-                    if (null !== $type = array_get($schema, $key)) {
-                        $result[$key] = DataFormatter::formatValue($value, $type);
+                } else {
+                    if (null !== $type = array_get($schema, $rkey)) {
+                        $result[$rkey] = DataFormatter::formatValue($row, $type);
                     }
                 }
             }
@@ -398,13 +392,13 @@ class StoredProcedure extends BaseDbResource
         $apis = [
             $path . '/{procedure_name}' => [
                 'get'  => [
-                    'tags'              => [$serviceName],
-                    'summary'           => 'call' . $capitalized . 'StoredProcedure() - Call a stored procedure.',
-                    'operationId'       => 'call' . $capitalized . 'StoredProcedure',
-                    'description'       =>
+                    'tags'        => [$serviceName],
+                    'summary'     => 'call' . $capitalized . 'StoredProcedure() - Call a stored procedure.',
+                    'operationId' => 'call' . $capitalized . 'StoredProcedure',
+                    'description' =>
                         'Call a stored procedure with no parameters. ' .
                         'Set an optional wrapper for the returned data set. ',
-                    'parameters'        => [
+                    'parameters'  => [
                         [
                             'name'        => 'procedure_name',
                             'description' => 'Name of the stored procedure to call.',
@@ -427,7 +421,7 @@ class StoredProcedure extends BaseDbResource
                             'required'    => false,
                         ],
                     ],
-                    'responses'         => [
+                    'responses'   => [
                         '200'     => [
                             'description' => 'Success',
                             'schema'      => ['$ref' => '#/definitions/StoredProcedureResponse']
@@ -439,15 +433,15 @@ class StoredProcedure extends BaseDbResource
                     ],
                 ],
                 'post' => [
-                    'tags'              => [$serviceName],
-                    'summary'           => 'call' .
+                    'tags'        => [$serviceName],
+                    'summary'     => 'call' .
                         $capitalized .
                         'StoredProcedureWithParams() - Call a stored procedure.',
-                    'operationId'       => 'call' . $capitalized . 'StoredProcedureWithParams',
-                    'description'       =>
+                    'operationId' => 'call' . $capitalized . 'StoredProcedureWithParams',
+                    'description' =>
                         'Call a stored procedure with parameters. ' .
                         'Set an optional wrapper and schema for the returned data set. ',
-                    'parameters'        => [
+                    'parameters'  => [
                         [
                             'name'        => 'procedure_name',
                             'description' => 'Name of the stored procedure to call.',
@@ -477,7 +471,7 @@ class StoredProcedure extends BaseDbResource
                             'required'    => false,
                         ],
                     ],
-                    'responses'         => [
+                    'responses'   => [
                         '200'     => [
                             'description' => 'Success',
                             'schema'      => ['$ref' => '#/definitions/StoredProcedureResponse']